Deployment

+++ utils/plotting_results.py
@@ -5,15 +5,18 @@ from PIL import Image
 
 from reportlab.pdfgen import canvas
 from reportlab.lib.pagesizes import landscape
+from reportlab.lib import colors
 
 from utils.crop_v2 import smart_cropping
 
 
 def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict, group_name2chosen_combinations,
-                           layouts_df,logger, gallery_path, output_save_path):
+                           layouts_df,logger, gallery_path, output_save_path,relation_type):
     gal_num = gallery_path.split('\\')[-1]
+    save_folder = os.path.join(output_save_path,gal_num)
+    os.makedirs(save_folder, exist_ok=True)
     # Create a PDF canvas with landscape orientation and A4 size (1:2 aspect ratio)
-    c = canvas.Canvas(os.path.join(output_save_path, f'{gal_num}.pdf'), pagesize=landscape((2000, 1000)))
+    c = canvas.Canvas(os.path.join(save_folder, f'{relation_type}.pdf'), pagesize=landscape((2000, 1000)))
     page_width, page_height = landscape((2000, 1000))
 
     if cover_img_layout_id is not None:
@@ -22,43 +25,64 @@ def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict
         cover_layout_info = ast.literal_eval(layouts_df.loc[int(cover_img_layout_id)]['boxes_info'])
         cover_image_id = cover_img['image_id'].values[0]
         cover_img_path = os.path.join(gallery_path, str(cover_image_id) + '.jpg')
-        cover_img_reading = Image.open(cover_img_path)
-        # Resize the image to fit the box
-        centroid = cover_img['background_centroid'].values[0]
-        for box in cover_layout_info:
-            x, y, w, h = box['x'], box['y'], box['width'], box['height']
 
-            # Adjust coordinates and dimensions to fit 1:2 aspect ratio
-            x = x * page_width
-            y = y * page_height
-            w = w * page_width
-            h = h * page_height
+        if os.path.exists(cover_img_path):
 
-            box_aspect_ratio = w / h
-            cropped_x, cropped_y, cropped_w, cropped_h = smart_cropping(float(cover_img['image_as'].iloc[0]),
-                                                                        cover_img['faces_info'], centroid,
-                                                                        float(cover_img['diameter'].iloc[0]),box_aspect_ratio)
-            # img.thumbnail((cropped_w, cropped_h))
-            np_img = np.array(cover_img_reading)
-            im_x = int(cropped_x * np_img.shape[0])
-            im_y = int(cropped_y * np_img.shape[1])
-            im_h = int(cropped_h * np_img.shape[0])
-            im_w = int(cropped_w * np_img.shape[1])
+            img = Image.open(cover_img_path)
+            np_img = np.array(img)
 
-            cropped_image = np_img[im_x:im_x + im_h, im_y:im_y + im_w]
+            # Resize the image to fit the box
+            centroid = cover_img['background_centroid'].values[0]
+            for box in cover_layout_info:
+                x, y, w, h = box['x'], box['y'], box['width'], box['height']
 
-            # Resize the cropped image
-            img = Image.fromarray(cropped_image)
+                # Adjust coordinates and dimensions to fit 1:2 aspect ratio
+                x = x * page_width
+                y = y * page_height
+                w = w * page_width
+                h = h * page_height
 
-            # Save the resized image to a temporary file
-            temp_img_path = "cover_temp.jpg"
-            img.save(temp_img_path)
+                # Add text above the image
+                text_x = x
+                text_y = y + h + 20  # Adjust the 20 to move the text further above the image if needed
 
-            # Plot the cover image inside the box
-            c.drawImage(temp_img_path, x, y, w, h)
+                # Set font and draw the text
+                c.setFont("Helvetica", 24)  # Set font and size
+                c.setFillColor(colors.black)  # Set text color
+                #c.drawString(text_x, text_y, f"{cover_img['image_orderInScene'].iloc[0]}")  # Customize the text as needed
 
-            # Clean up the temporary image file
-            os.remove(temp_img_path)
+                box_aspect_ratio = w / h
+                faces = np.copy(cover_img['faces_info'])
+                cropped_x, cropped_y, cropped_w, cropped_h = smart_cropping(float(cover_img['image_as'].iloc[0]),
+                                                                            faces, centroid,
+                                                                            float(cover_img['diameter'].iloc[0]),box_aspect_ratio)
+                # Denormalize the coordinates based on the image dimensions
+                im_x = int(cropped_x * np_img.shape[1])  # Multiply by image width
+                im_y = int(cropped_y * np_img.shape[0])  # Multiply by image height
+                im_w = int(cropped_w * np_img.shape[1])  # Multiply by image width
+                im_h = int(cropped_h * np_img.shape[0])  # Multiply by image height
+
+                # Ensure the coordinates are within image boundaries
+                im_x = max(0, min(im_x, np_img.shape[1] - 1))
+                im_y = max(0, min(im_y, np_img.shape[0] - 1))
+                im_w = max(0, min(im_w, np_img.shape[1] - im_x))
+                im_h = max(0, min(im_h, np_img.shape[0] - im_y))
+
+                # Crop the image using correct indexing
+                cropped_image = np_img[im_y:im_y + im_h, im_x:im_x + im_w]
+
+                # Resize the cropped image
+                img = Image.fromarray(cropped_image)
+
+                # Save the resized image to a temporary file
+                temp_img_path = "cover_temp.jpg"
+                img.save(temp_img_path)
+
+                # Plot the cover image inside the box
+                c.drawImage(temp_img_path, x, y, w, h)
+
+                # Clean up the temporary image file
+                os.remove(temp_img_path)
 
     for group_name in sorted_sub_groups_dict.keys():
             if group_name not in group_name2chosen_combinations.keys():
@@ -100,6 +124,8 @@ def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict
 
                     c_image_info = c_group[c_group['image_id'] == c_image_id]
                     img_path = os.path.join(gallery_path, str(cur_photo.id) + '.jpg')
+                    if not os.path.exists(img_path):
+                        continue
                     img = Image.open(img_path)
 
                     x, y, w, h = box['x'], box['y'], box['width'], box['height']
@@ -110,6 +136,17 @@ def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict
                     w = w * page_width
                     h = h * page_height
 
+                    # Add text above the image
+                    text_x = x
+                    text_y = y + h + 20  # Adjust the 20 to move the text further above the image if needed
+
+                    # Set font and draw the text
+                    c.setFont("Helvetica", 24)  # Set font and size
+                    c.setFillColor(colors.black)  # Set text color
+                    #c.drawString(text_x, text_y,
+                                 #f"{c_image_info['image_orderInScene'].iloc[0]}")  # Customize the text as needed
+
+
                     box_aspect_ratio = w / h
                     centroid = c_image_info['background_centroid'].values[0]
                     cropped_x, cropped_y, cropped_w, cropped_h = smart_cropping(float(c_image_info['image_as'].iloc[0]),
@@ -119,12 +156,20 @@ def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict
 
                     np_img = np.array(img)
 
-                    im_x = int(cropped_x * np_img.shape[0])
-                    im_y = int(cropped_y * np_img.shape[1])
-                    im_h = int(cropped_h * np_img.shape[0])
-                    im_w = int(cropped_w * np_img.shape[1])
+                    # Denormalize the coordinates based on the image dimensions
+                    im_x = int(cropped_x * np_img.shape[1])  # Multiply by image width
+                    im_y = int(cropped_y * np_img.shape[0])  # Multiply by image height
+                    im_w = int(cropped_w * np_img.shape[1])  # Multiply by image width
+                    im_h = int(cropped_h * np_img.shape[0])  # Multiply by image height
 
-                    cropped_image = np_img[im_x:im_x + im_h, im_y:im_y + im_w]
+                    # Ensure the coordinates are within image boundaries
+                    im_x = max(0, min(im_x, np_img.shape[1] - 1))
+                    im_y = max(0, min(im_y, np_img.shape[0] - 1))
+                    im_w = max(0, min(im_w, np_img.shape[1] - im_x))
+                    im_h = max(0, min(im_h, np_img.shape[0] - im_y))
+
+                    # Crop the image using correct indexing
+                    cropped_image = np_img[im_y:im_y + im_h, im_x:im_x + im_w]
                     # Resize the cropped image
                     img = Image.fromarray(cropped_image)
 
@@ -141,7 +186,8 @@ def plot_album(cover_img, cover_img_layout_id,sub_groups, sorted_sub_groups_dict
 
                 # Add spread number as a label (optional)
                 c.setFont("Helvetica", 12)
-                c.drawString(30, 30, f"Cluster {group_name} - Spread {spread_index + 1}")
+                #c.drawString(30, 30, f"Cluster {group_name} - Spread {spread_index + 1}")
+
 
     # Save the PDF
     c.save()
